Update microsoft-word-quiz.md (#983)
* Update microsoft-word-quiz.md

Answered Inline Comment

* Update microsoft-word-quiz.md

Q10:
Right-click a table and choose a new style. <- Is this new with Office356?
Apply a table style. <- This would be the way in Office2019

* Update microsoft-word-quiz.md

No solution corresponds to the functions in Office2019. Found solutions added as comment box

* Update microsoft-word-quiz.md

added personal hint

* Update microsoft-word-quiz.md

Repaired Images Links

* Update microsoft-word-quiz.md

Text formatting of given answer suggestions formatted

* Update microsoft-word-quiz.md

#### Q38. You need to add a trademark symbol to a new product name. How can you accomplish this? (Select two items.) 
I'm not sure, but i could not find the menchend and selected **Icons** in the Illustration group in the Insert tab. 
Since (R) is also a trademark sign I think this is the better solution

* Update microsoft-word-quiz.md

Answer options selected for question 40 and 41

* Update microsoft-word-quiz.md